// src/WebUI/ClientApp/src/app/lecacy-app/core/models/elaraby/IndexWorkOrder.ts
export class IndexWorkOrder {
  OrderNumber: string;
  CustomerName: string;
  AreaName: string;
  LandLine: string;
  Address: string;
  ProductType: string;
  ProblemCliet: string;
  MobileNumber: string;
  ScheduledStartDate: string;
  BackOfficeNotes: string;
  WorkOrderType: string;
  HasComplaint: string;
  SapNumber: string;
  WorkOrderId: string;
  Oid: string;
}